+++ dask_match/collection/__init__.py
@@ -1,0 +1,1 @@
+from dask_match.collection.core import *
