+++ dask_match/expr/__init__.py
@@ -1,0 +1,2 @@
+from dask_match.expr.core import *
+from dask_match.expr.io import *
