+++ dask_match/expr/io/parquet.py
@@ -14,8 +14,9 @@ from dask.utils import natural_sort_key
 from fsspec.utils import stringify_path
 from matchpy import CustomConstraint, Pattern, ReplacementRule, Wildcard
 
-from dask_match.core import EQ, GE, GT, LE, LT, NE, Filter, new_collection
-from dask_match.io import IO
+from dask_match.collection.core import new_collection
+from dask_match.expr.core import EQ, GE, GT, LE, LT, NE, Filter
+from dask_match.expr.io import IO
 
 NONE_LABEL = "__null_dask_index__"
