+++ dask_match/expr/reductions.py
@@ -4,7 +4,7 @@ from dask.dataframe.core import _concat, is_series_like
 from dask.utils import M, apply
 from matchpy import Pattern, ReplacementRule, Wildcard
 
-from dask_match.core import Expr
+from dask_match.expr.core import Expr
 
 
 class ApplyConcatApply(Expr):
